Update README.md
This is a conceptual system of how a traditional Payment system should be designed using DDD, EventStore for persisting events and MongoDB for the read model.

The Application services are responsible for issuing one-way commands to the system. Commands are responsible for raising appropriate aggregates. Aggregates should encapsulate their own domain events. An important part of the system is the reactors. These are different kind of subscription services, which apart from the projections are also responsible to act upon the domain events.